refactor: isolate ClickClack binding retention (#119358)
